refactor: refactor error handling code to generic function
